% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{add_row_id}
\alias{add_row_id}
\title{Add a row ID column to a data frame}
\usage{
add_row_id(data, name = ".id")
}
\arguments{
\item{data}{A data frame to which the row ID column will be added.}

\item{name}{The name of the new row ID column. Default is ".id".}
}
\value{
The input data frame with an additional column containing row IDs.
}
\description{
This function adds a new column to a data frame containing a unique identifier for each row. The default column name is ".id", but this can be customized using the \code{name} parameter.
}
\examples{
df <- data.frame(x = c(1, 2, 3), y = c("a", "b", "c"))
add_row_id(df)
}
