% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_all_unique}
\alias{cols_all_unique}
\title{Identify columns with all unique values}
\usage{
cols_all_unique(data)
}
\arguments{
\item{data}{The data frame}
}
\value{
A character vector of column names with all unique values.
}
\description{
Find columns that have all unique values. If there are no identified columns, it will return NA.
}
\examples{
cols_all_unique(ChickWeight)

}
\seealso{
Other quality checks:
\code{\link[=cols_bijective]{cols_bijective()}},
\code{\link[=cols_constant]{cols_constant()}},
\code{\link[=cols_identify_all]{cols_identify_all()}},
\code{\link[=cols_missing]{cols_missing()}},
\code{\link[=cols_nested]{cols_nested()}}
}
\concept{quality checks}
