% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_constant}
\alias{cols_constant}
\alias{cols_unary}
\title{Identify constant columns}
\usage{
cols_constant(data, na.rm = FALSE)

cols_unary(data, na.rm = FALSE)
}
\arguments{
\item{data}{The data frame}

\item{na.rm}{Should missing values be removed?}
}
\value{
A character vector of constant columns.
}
\description{
Similar to janitor::remove_constant, except this function
aims to identify the constant columns without removing them.
If there are no constant columns, it will return NA.
}
\examples{
df <- data.frame(x1 = 1, x2 = c("A", "B", "C"), x3 = "a")
cols_constant(df)
}
\seealso{
Other quality checks:
\code{\link[=cols_all_unique]{cols_all_unique()}},
\code{\link[=cols_bijective]{cols_bijective()}},
\code{\link[=cols_identify_all]{cols_identify_all()}},
\code{\link[=cols_missing]{cols_missing()}},
\code{\link[=cols_nested]{cols_nested()}}
}
\concept{quality checks}
