% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_missing}
\alias{cols_missing}
\title{Identify columns that have certain proportion of missing values}
\usage{
cols_missing(data, cutoff = 1)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values. The
default of 1 finds columns that are entirely missing. A cutoff of 0 finds
columns with any missing values.}
}
\value{
A character vector of column names with missing values with at least a certain proportion.
}
\description{
Similar to janitor::remove_empty, except this function
aims to identify the columns without removing them.
If there are no identified columns, it will return NA.
}
\examples{
# find columns that have any missing values
cols_missing(airquality, 0)
}
\seealso{
Other quality checks:
\code{\link[=cols_all_unique]{cols_all_unique()}},
\code{\link[=cols_bijective]{cols_bijective()}},
\code{\link[=cols_constant]{cols_constant()}},
\code{\link[=cols_identify_all]{cols_identify_all()}},
\code{\link[=cols_nested]{cols_nested()}}
}
\concept{quality checks}
