% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_binary}
\alias{cols_binary}
\alias{cols_ternary}
\alias{cols_multinary}
\title{Identify columns with a specific number of unique values}
\usage{
cols_binary(data, na.rm = FALSE)

cols_ternary(data, na.rm = FALSE)

cols_multinary(data, na.rm = FALSE, n = 1)
}
\arguments{
\item{data}{The data frame}

\item{na.rm}{Logical, whether to remove NA values before counting unique values.}

\item{n}{The specific number of unique values to look for.}
}
\value{
A character vector of column names that have exactly n unique values.
}
\description{
This function identifies columns in a data frame that have a specific number of unique values. It can be used to find constant, binary, ternary, or any other n-ary columns. If there are no identified columns, it will return NA.
}
\examples{
cols_binary(mtcars)
}
\concept{quality-checks}
