% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{cols_nested}
\alias{cols_nested}
\title{Nestedness checks}
\usage{
cols_nested(data, na.rm = FALSE)
}
\arguments{
\item{data}{The data frame to be checked for nestedness.}

\item{na.rm}{Remove NA values when checking for nestedness.}
}
\value{
A list of identified nested columns. The first element of each list is the child variable, and the second element is the parent variable.
}
\description{
This function checks for nestedness in a data frame. It identifies columns that are nested within other columns, meaning that the values in one column are subsets of the values in another column. The function returns a list of identified nested columns and can print the results to the console.
}
\examples{
cols_nested(CO2)
}
\seealso{
Other quality checks:
\code{\link[=cols_all_unique]{cols_all_unique()}},
\code{\link[=cols_bijective]{cols_bijective()}},
\code{\link[=cols_constant]{cols_constant()}},
\code{\link[=cols_identify_all]{cols_identify_all()}},
\code{\link[=cols_missing]{cols_missing()}}
}
\concept{quality checks}
