% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Render EDA report}
\usage{
create_report(.data, ..., .regroup = TRUE,
  output_format = html_document(toc = TRUE, toc_depth = 2),
  output_file = "report.html", output_dir = getwd(),
  options = report_options())
}
\arguments{
\item{.data}{a data frame}

\item{...}{unquoted columns names of variables to be considered as response}

\item{.regroup}{logical. Should the groups with less than 5 occurences be grouped together.}

\item{output_format}{see \link{render}}

\item{output_file}{see \link{render}}

\item{output_dir}{see \link{render}}

\item{options}{list. Options for report}
}
\value{
a report at the format specified in output_format
}
\description{
Render EDA report
}
\examples{
##ToDo
}
