% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascan-package.R
\docType{package}
\name{datascan-package}
\alias{datascan}
\alias{datascan-package}
\title{datascan: Scan data for quick structural summaries and checks}
\description{
Scans data for checking columns that are constant, bijective, missing or all unique. Users can also try to identify the columns that uniquely indexes the observational unit and whether some columns are nested or crossed.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
