% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract}
\alias{extract}
\title{Extract operator}
\usage{
x |> extract(i, j)
}
\value{
The subset of \code{x} selected by \code{i} and \code{j}, equivalent to \code{x[i, j]}.
}
\description{
See \code{magrittr::extract} for details.
This works well in conjunction with the concurrence_matrix function to extract specific rows and columns from the resulting matrix.
}
\keyword{internal}
