% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{is_complete}
\alias{is_complete}
\title{Is variable complete within another variable?}
\usage{
is_complete(x, y, na.rm = TRUE)
}
\arguments{
\item{x, y}{Vectors of the same size where \code{x} (child variable) is checked for completeness within \code{y} (parent variable)}

\item{na.rm}{Remove NA values when checking for completeness.}
}
\value{
TRUE if the values in vector x are complete
}
\description{
This function checks if a variable is complete within another variable. If all levels of the first variable appear in all levels of the second variable, it returns TRUE; otherwise, it returns FALSE. The function can also handle NA values based on the \code{na.rm} parameter.
}
\examples{
df <- data.frame(
  site = c("A", "A", "B", "B"),
  plot1 = c("A1", "A2", "B1", "B2"),
  plot2 = c("A1", "A2", "A1", "A2")
)
is_complete(df$plot1, df$site) # FALSE
is_complete(df$plot2, df$site) # TRUE
}
