% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical-checks.R
\name{is_numerical}
\alias{is_numerical}
\title{Is the variable numerical?}
\usage{
is_numerical(x)
}
\arguments{
\item{x}{A vector to be checked for numericality.}
}
\value{
TRUE if the variable is numerical, FALSE otherwise.
}
\description{
This function checks if a variable is numerical. It returns TRUE if the variable is numerical and FALSE otherwise. If non-numerical entries are found, it prints a message indicating the non-numerical entries.
}
\examples{
is_numerical(c("1", "2", "3"))
is_numerical(c("1", "2", "3", "A"))
}
\seealso{
Other numerical:
\code{\link[=as_numerical]{as_numerical()}}
}
\concept{numerical}
