% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{is_observational_unit}
\alias{is_observational_unit}
\alias{is_obs_unit}
\title{Is it the observational unit?}
\usage{
is_observational_unit(.data, ..., .check_redundancy = TRUE)

is_obs_unit(.data, ..., .check_redundancy = TRUE)
}
\arguments{
\item{.data}{The data frame}

\item{...}{The columns in the data frame that potentially index the observational unit.}

\item{.check_redundancy}{Logical. If TRUE, checks if any of the columns are redundant.}
}
\value{
A logical value.
}
\description{
An observational unit is the entity that is being measured in a study. For example, if blood pressure of a person is measured, then the person is the observational unit. If blood pressure is measured multiple times for a person, then the observational unit is the person and the time of measurement.
}
\details{
Supply the columns that potentially index the observational unit. The column cannot uniquely identify the row. If the number of rows in the data frame is equal to the number of distinct rows in the selected columns, then it returns TRUE, otherwise FALSE. If TRUE, it checks if any of the columns are redundant.
}
\examples{
is_observational_unit(ChickWeight, Time, Chick)
is_observational_unit(ChickWeight, Time, Chick, Diet)

}
