% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar plot of a categorical column in a dataframe}
\usage{
plot_bar(.data, .cat, .by, nas = TRUE)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat}{unquoted name of a categrocial column}

\item{.by}{(optional) unquoted name of a categrocial column}

\item{nas}{logical. Should missing values be considered as group on x-axis}
}
\value{
a ggplot
}
\description{
Bar plot of a categorical column in a dataframe
}
\examples{
library(dplyr)
plot_bar(starwars, eye_color)
}
