% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{plot_count}
\alias{plot_count}
\title{Count overlapping points of two categorical columns in a dataframe}
\usage{
plot_count(.data, .cat1, .cat2)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat1}{a categorical column}

\item{.cat2}{a categorical column}
}
\value{
a ggplot
}
\description{
Count overlapping points of two categorical columns in a dataframe
}
\examples{
library(dplyr)
plot_count(starwars, eye_color, hair_color)
}
