% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot between two numerical columns of a data frame}
\usage{
plot_scatter(.data, .num1, .num2, .by = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{.num1}{unquoted name of numercial column for x-axis}

\item{.num2}{unquoted name of numercial column for y-axis}

\item{.by}{default \code{NULL}. Unqoted name of categorical column to color points by groups}
}
\value{
a ggplot
}
\description{
Scatter plot between two numerical columns of a data frame
}
\examples{
# With .by = NULL
plot_scatter(iris, Sepal.Length, Sepal.Width)

# With color by groups
plot_scatter(iris, Sepal.Length, Sepal.Width, .by = Species)
}
