% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_options.R
\name{report_options}
\alias{report_options}
\title{Config options for automatic report}
\usage{
report_options(
  .title = "Exploratory Data Analysis report",
  .author = "John Doe",
  .date = Sys.Date()
)
}
\arguments{
\item{.title}{report title}

\item{.author}{report author}

\item{.date}{report date}
}
\value{
a list of config values
}
\description{
Config options for automatic report
}
