% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{rows_missing}
\alias{rows_missing}
\title{Identify rows that have certain proportion of missing values}
\usage{
rows_missing(data, cutoff = 1)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values. The
default of 1 finds rows that are entirely missing. A cutoff of 0 finds
rows with any missing values.}
}
\value{
An integer vector of rows with missing values with at least a certain proportion.
}
\description{
Similar to janitor::remove_empty, except this function
aims to identify the columns without removing them.
If there are no identified columns, it will return NA.
}
\examples{
rows_missing(airquality, 0.1)

}
