% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{Render EDA report}
\usage{
scan(.data, ..., .regroup = TRUE, output_format = html_document(toc =
  TRUE, toc_depth = 3, theme = "journal"), output_file = "report.html",
  output_dir = getwd(), clean = FALSE)
}
\arguments{
\item{.data}{a data frame}

\item{...}{unquoted columns names of variables to be considered as response}

\item{.regroup}{logical. Should the groups with less than 5 occurences be grouped together.}

\item{output_format}{see \link{render}}

\item{output_file}{see \link{render}}

\item{output_dir}{see \link{render}}

\item{clean}{see \link{render}}
}
\value{
a report at the format specified in output_format
}
\description{
Render EDA report
}
\examples{
##ToDo
}
