% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_columns.R
\name{scan_columns}
\alias{scan_columns}
\title{Summary values about columns in a dataframe}
\usage{
scan_columns(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a dataframe with the summary values
}
\description{
Summary values about columns in a dataframe
}
\examples{
library(dplyr)
scan_columns(starwars)
}
