% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_corr.R
\name{scan_corr}
\alias{scan_corr}
\title{Correlation between numerical columns of a dataframe}
\usage{
scan_corr(.data, .use = "na.or.complete", .method = "spearman")
}
\arguments{
\item{.data}{a dataframe}

\item{.use}{see \link[stats]{cor}}

\item{.method}{see \link[stats]{cor}}
}
\value{
a data frame with correlation values of all pairs of numerical columns
}
\description{
Correlation between numerical columns of a dataframe
}
\examples{
scan_corr(iris)
}
