% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_data.R
\name{scan_data}
\alias{scan_data}
\title{General summary values about a dataframe}
\usage{
scan_data(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a datframe with the summary values
}
\description{
General summary values about a dataframe
}
\examples{
scan_data(iris)
}
