% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_groups.R
\name{scan_groups}
\alias{scan_groups}
\title{Frequency of groups in all categorical columns of a dataframe}
\usage{
scan_groups(.data)
}
\arguments{
\item{.data}{a data.frame}
}
\value{
a data.frame with one row for each group in each categorical variables
in \code{data} with \code{variables}, the categorical varibales names, \code{groups}
the groups names of categorical variables, \code{n} the occurence number for each
groups and \code{p} the proportion
}
\description{
Frequency of groups in all categorical columns of a dataframe
}
\examples{
data(iris)
scan_groups(iris)
}
