% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_ngcovar.R
\name{scan_ngcovar}
\alias{scan_ngcovar}
\title{Statistics of numerical columns by groups of catagorical columns of a dataframe}
\usage{
scan_ngcovar(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a data frame with identifiers for numerical and categorical (and groups) columns
and the same summary values as \link[datascan]{scan_numerics}
}
\description{
Statistics of numerical columns by groups of catagorical columns of a dataframe
}
\examples{
scan_ngcovar(iris)
}
