% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_numbygrp.R
\name{scan_numbygrp}
\alias{scan_numbygrp}
\title{Statistics of numerics columns by groups of catagorical columns}
\usage{
scan_numbygrp(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a data frame with identifiers for numerics and categorical (and groups) columns
and the same summary values as \link[datascan]{scan_numerics}
}
\description{
Statistics of numerics columns by groups of catagorical columns
}
\examples{
scan_numbygrp(iris)
}
