% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_r.R
\name{scan_r}
\alias{scan_r}
\title{Get the R coefficient of linear models between numeric and catagorical variables}
\usage{
scan_r(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a data frame with paired names of categorical variables (\code{Cat}
and mnumeric ones \code{Num} and \code{r} value)
}
\description{
Get the R coefficient of linear models between numeric and catagorical variables
}
\examples{
vis_r(dplyr::starwars)
}
