% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-style.R
\name{utils-style}
\alias{utils-style}
\alias{.prop}
\alias{.color_bar}
\alias{.apply_colorbar}
\alias{.nonzero}
\title{Some internals functions}
\usage{
.prop(x, nr)

.color_bar(x, nr, color = "lightgreen", simple = TRUE)

.apply_colorbar(x, nr)

.nonzero(x)
}
\arguments{
\item{x}{a numeric vector}

\item{nr}{a numeric value}

\item{color}{css color}

\item{simple}{a logical}
}
\description{
Some internals functions
}
