% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.get_madm}
\alias{.get_skewness}
\alias{.get_bins}
\alias{.get_cramerv}
\alias{.get_r}
\alias{.transform_cat}
\alias{.clean_names}
\title{Some internals functions}
\usage{
.get_madm(.num)

.get_skewness(.num)

.get_bins(.num)

.get_cramerv(.cat1, .cat2)

.get_r(.data, .cat, .num)

.transform_cat(.cat)

.clean_names(.data)
}
\arguments{
\item{.num}{a numeric vector}

\item{.cat1}{a categorical vector}

\item{.cat2}{a categorical vector}

\item{.data}{a data frame}

\item{.cat}{a categorical vector}
}
\description{
Some internals functions
}
\examples{
.get_madm(rnorm(100))
.get_skewness(rnorm(100))
.get_bins(rnorm(100))
x <- sample(letters, 100, replace = TRUE)
.get_cramerv(x, x)
.get_r(iris, "Species", "Sepal.Length")
library(dplyr)
.transform_cat(starwars$species)
.transform_cat(starwars$gender)
dt <- tibble::tibble('a/b' = rnorm(10))
.clean_names(dt)
}
