% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_corr.R
\name{vis_corr}
\alias{vis_corr}
\title{plots correlation between numerical columns of a dataframe}
\usage{
vis_corr(.data, .use = "pairwise.complete.obs", .method = "spearman")
}
\arguments{
\item{.data}{a dataframe}

\item{.use}{see \link[stats]{cor}}

\item{.method}{see \link[stats]{cor}}
}
\value{
a ggplot
}
\description{
plots correlation between numerical columns of a dataframe
}
\examples{
vis_corr(iris)
}
