% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cramerv.R
\name{vis_cramerv}
\alias{vis_cramerv}
\title{Plot bias corrected Cramer's V for all pairs of categorical columns}
\usage{
vis_cramerv(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a ggplot
}
\description{
Plot bias corrected Cramer's V for all pairs of categorical columns
}
\examples{
vis_cramerv(dplyr::starwars)
}
