% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ggcovar.R
\name{vis_ggcovar}
\alias{vis_ggcovar}
\title{Visualise frequency of paired group between categorical variables in a data frame}
\usage{
vis_ggcovar(.data, ..., .regroup = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{...}{you can use this arguement to specify targeted categorical columns that will be used as y-axis variable}

\item{.regroup}{if \code{TRUE}, groups with less than 5 occurences are grouped together under the name "Others_groups"}
}
\value{
a list of ggplot
}
\description{
Visualise frequency of paired group between categorical variables in a data frame
}
\examples{
\dontrun{
library(dplyr)
# By regrouping small groups
vis_ggcovar(starwars)

# Without regrouping small groups
vis_ggcovar(starwars, .regroup = FALSE)

# With a target column
vis_ggcovar(starwars, skin_color)
}
}
