% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_groups.R
\name{vis_groups}
\alias{vis_groups}
\title{Get and Visualise distribution of a numeric vector}
\usage{
vis_groups(.cat, .labx = "", .plot = TRUE)
}
\arguments{
\item{.cat}{a categorical vector}

\item{.labx}{label of x-axix}

\item{.plot}{if \code{TRUE} the plot is displayed}
}
\value{
a ggplot object
}
\description{
Get and Visualise distribution of a numeric vector
}
\examples{
vis_groups(iris[,"Species"])
}
