% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_groupsbygroups.R
\name{vis_groupsbygroups}
\alias{vis_groupsbygroups}
\title{Visualise frequency of paired group between categorical variables in a data frame}
\usage{
vis_groupsbygroups(.data, .regroup = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{.regroup}{if \code{TRUE}, groups with less than 5 occurences are grouped together under the name "Others_groups"}
}
\value{
a list of ggplot
}
\description{
Visualise frequency of paired group between categorical variables in a data frame
}
\examples{
library(dplyr)
vis_groupsbygroups(starwars)
}
