% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_grp.R
\name{vis_grp}
\alias{vis_grp}
\title{Ggplot of group frequency in a categorical variable}
\usage{
vis_grp(.cat, .labx = "")
}
\arguments{
\item{.cat}{a categorical vector}

\item{.labx}{labels of x-axis}
}
\value{
a ggplot
}
\description{
Ggplot of group frequency in a categorical variable
}
\examples{
library(dplyr)
starwars \%>\%
 pull(eye_color) \%>\%
 vis_grp()
}
