% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_grpbygrp.R
\name{vis_grpbygrp}
\alias{vis_grpbygrp}
\title{Ggplot of frequency of parired group between two categorical variables}
\usage{
vis_grpbygrp(.cat1, .cat2, .labx = "", .laby = "")
}
\arguments{
\item{.cat1}{a categorical vector}

\item{.cat2}{a categorical vector}

\item{.labx}{label for x-axis}

\item{.laby}{label for y-axis}
}
\value{
a ggplot
}
\description{
Ggplot of frequency of parired group between two categorical variables
}
\examples{
library(dplyr)
with(starwars, vis_grpbygrp(eye_color, hair_color))
}
