% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ngcovar.R
\name{vis_ngcovar}
\alias{vis_ngcovar}
\title{Violin plots for all pairs of numerical/categorical columns in a data frame}
\usage{
vis_ngcovar(.data, ..., .regroup = TRUE)
}
\arguments{
\item{.data}{a dataframe}

\item{...}{you can use this arguement to specify targeted numerical columns that will be used as y-axis variable}

\item{.regroup}{if \code{TRUE}, groups with less than 5 occurences are grouped together under the name "Others_groups"}
}
\value{
a list of ggplot
}
\description{
Violin plots for all pairs of numerical/categorical columns in a data frame
}
\examples{
\dontrun{
library(dplyr)
# By regrouping small groups
vis_ngcovar(starwars)

# Without regrouping small groups
vis_ngcovar(starwars, .regroup = FALSE)

# With a target column
vis_ngcovar(starwars, height)
}
}
