% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nncovar.R
\name{vis_nncovar}
\alias{vis_nncovar}
\title{Scatter plots of all pairs of numerical variables in a data.frame}
\usage{
vis_nncovar(.data, ..., .by = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{...}{you can use this arguement to specify targeted numerical columns that will be used as y-axis variable}

\item{.by}{color scatter by groups}
}
\value{
a list of ggplots
}
\description{
Scatter plots of all pairs of numerical variables in a data.frame
}
\examples{
# All combinations
vis_nncovar(iris)

# With a targeted y-axis variable
vis_nncovar(iris, Sepal.Length)
}
