% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_numbygrp.R
\name{vis_numbygrp}
\alias{vis_numbygrp}
\title{Violin plot of a numeric vector by groups of a categorical vectors}
\usage{
vis_numbygrp(.cat, .num, .labx = "", .laby = "")
}
\arguments{
\item{.cat}{a catagorical vector}

\item{.num}{a numeric vector}

\item{.labx}{label for x-axis}

\item{.laby}{label for y-axis}
}
\value{
a ggplot
}
\description{
Violin plot of a numeric vector by groups of a categorical vectors
}
\examples{
vis_numbygrp(iris$Species, iris$Sepal.Length)
}
