% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_numerics.R
\name{vis_numerics}
\alias{vis_numerics}
\title{Get and Visualise distribution of a numeric vector}
\usage{
vis_numerics(.num, .labx = "", .bins = .get_bins(.num), .plot = TRUE)
}
\arguments{
\item{.num}{a numeric vector}

\item{.labx}{label of x-axix}

\item{.bins}{bins argument for \code{geom_histogram}}

\item{.plot}{if \code{TRUE} the plot is displayed}
}
\value{
a ggplot object
}
\description{
Get and Visualise distribution of a numeric vector
}
\examples{
vis_numerics(rnorm(100))
}
