% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_numerics_all.R
\name{vis_numerics_all}
\alias{vis_numerics_all}
\title{Get and Visualise distribution of all numeric column in a data frame}
\usage{
vis_numerics_all(.data, .plot = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{.plot}{if \code{TRUE} the plot is displayed}
}
\value{
a list of ggplot objects
}
\description{
Get and Visualise distribution of all numeric column in a data frame
}
\examples{
vis_numerics_all(iris)
}
