% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_numericsbygroups.R
\name{vis_numericsbygroups}
\alias{vis_numericsbygroups}
\alias{vis_nbg}
\title{Violin plots for all pairs of numeric/categorical variables in a data frame}
\usage{
vis_numericsbygroups(.data, .regroup = TRUE)

vis_nbg(.data)
}
\arguments{
\item{.data}{a dataframe}

\item{.regroup}{if \code{TRUE}, groups with less than 5 occurences are grouped together under the name "Others_groups"}
}
\value{
a list of ggplot
}
\description{
Violin plots for all pairs of numeric/categorical variables in a data frame
}
\examples{
library(dplyr)
vis_numericsbygroups(starwars)
}
