% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_r.R
\name{vis_r}
\alias{vis_r}
\title{Plot R coefficient of linear models between numeric and catagorical columns in a dataframe}
\usage{
vis_r(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a gpplot
}
\description{
Plot R coefficient of linear models between numeric and catagorical columns in a dataframe
}
\examples{
library(dplyr)
vis_r(starwars)
}
