% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_bijective}
\alias{cols_bijective}
\title{Identify bijective (one-to-one correspondence) columns}
\usage{
cols_bijective(data)
}
\arguments{
\item{data}{The data frame}
}
\value{
A list of character vectors of bijective columns.
}
\description{
Any values that have all unique values will not be included in the
output. If there are no identified columns, it will return an emtpy list.
}
\examples{
cols_bijective(OrchardSprays)

}
\seealso{
Other quality checks:
\code{\link[=cols_all_unique]{cols_all_unique()}},
\code{\link[=cols_constant]{cols_constant()}},
\code{\link[=cols_identify_all]{cols_identify_all()}},
\code{\link[=cols_missing]{cols_missing()}},
\code{\link[=cols_nested]{cols_nested()}}
}
\concept{quality checks}
