% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-checks.R
\name{cols_identify_all}
\alias{cols_identify_all}
\title{Identify all columns with potential issues}
\usage{
cols_identify_all(data, cutoff = 1, na.rm = FALSE, print = TRUE)
}
\arguments{
\item{data}{The data frame}

\item{cutoff}{The minimum cutoff for the proportion of missing values.}

\item{na.rm}{Should missing values be removed?}

\item{print}{Print the output or not.}
}
\value{
A named list with one element per check.
}
\description{
This function identifies all columns with potential issues, including all unique values, bijective columns, constant columns, and columns with missing values above a certain cutoff. It returns a list of identified columns for each category and can print the results to the console.
}
\examples{
cols_identify_all(airquality)

}
\seealso{
Other quality checks:
\code{\link[=cols_all_unique]{cols_all_unique()}},
\code{\link[=cols_bijective]{cols_bijective()}},
\code{\link[=cols_constant]{cols_constant()}},
\code{\link[=cols_missing]{cols_missing()}},
\code{\link[=cols_nested]{cols_nested()}}
}
\concept{quality checks}
