% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrence.R
\name{concurrence_matrix}
\alias{concurrence_matrix}
\alias{concurrence_table}
\title{Get the concurrence matrix or table}
\usage{
concurrence_matrix(data, x, group, na.rm = FALSE)

concurrence_table(data, x, group, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame containing the MET data}

\item{x}{Categorical variable to be used for the concurrence matrix.}

\item{group}{Categorical grouping variable.}
}
\value{
A concurrence matrix or table
}
\description{
Find the number of common levels of first variable across the levels of the second variable.
The diagonal shows the total number of levels at the corresponding level.
The off-diagonal shows the number of common levels between the two levels.
The matrix is symmetric.
}
\examples{
df <- expand.grid(gen = paste0("G", 1:20), env = paste0("E", 1:15))
df <- df[sample(nrow(df), 100), ]
(mat <- concurrence_matrix(df, gen, env))
extract(mat, "E1", "E2")
extract(mat, "E1", c("E2", "E3"))
extract(mat, c("E5", "E6", "E7"), c("E2", "E3"))
proportions(mat, 1) # proportion of levels in the first variable at each level of the second variable

}
