% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascan-package.R
\docType{package}
\name{datascan-package}
\alias{datascan}
\alias{datascan-package}
\title{datascan: Scan Data for Quick Structural Summaries and Checks}
\description{
Scans data for checking columns that are constant, one-to-one, missing or all unique. Users can also try to identify the columns that uniquely index the observational unit and whether some columns are nested or complete.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/datascan}
  \item \url{https://emitanaka.org/datascan/}
  \item Report bugs at \url{https://github.com/emitanaka/datascan/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
