% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar plot of a categorical column in a dataframe}
\usage{
plot_bar(
  .data,
  .cat,
  .by,
  nas = TRUE,
  type = "count",
  strwidth = 30,
  order = TRUE,
  autoflip = TRUE,
  maxp = 1,
  ystep = 5
)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat}{unquoted name of a categrocial column}

\item{.by}{(optional) unquoted name of a categrocial column}

\item{nas}{logical. Should missing values be considered as group on x-axis}

\item{type}{Either "count" or "percent" for y-axis}

\item{strwidth}{with of labels (for \code{stringr::str_wrap})}

\item{order}{(lgl) should the group be ordered by frequency? (chr) Order of .cat groups in barplot}

\item{autoflip}{(lgl) flip coordinates if there is more than 5 groups}

\item{maxp}{only if \code{type = "percent"}. Max percentage of y axis}

\item{ystep}{step for y-axis. \code{ystep} is multiplied by \code{10^(nchar(maxy) - 2)} where maxy is the max of y values}
}
\value{
a ggplot
}
\description{
Bar plot of a categorical column in a dataframe
}
\examples{
library(dplyr)
plot_bar(starwars, eye_color)
}
