% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Box plot of a numerical column by groups of a categorical column in a dataframe}
\usage{
plot_box(.data, .cat, .num, .by)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat}{a catagorical column name}

\item{.num}{a numeric numerical name}

\item{.by}{categorical columns names to add a second categorical column}
}
\value{
a ggplot
}
\description{
Box plot of a numerical column by groups of a categorical column in a dataframe
}
\examples{
plot_box(iris, Species, Sepal.Length)
}
