% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot of a numerical column by groups of a categorical column in a dataframe}
\usage{
plot_violin(.data, .cat, .num)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat}{a catagorical vector}

\item{.num}{a numeric vector}
}
\value{
a ggplot
}
\description{
Violin plot of a numerical column by groups of a categorical column in a dataframe
}
\examples{
plot_violin(iris, Species, Sepal.Length)
}
