% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot of a numerical column by groups of a categorical column in a dataframe}
\usage{
plot_violin(.data, .cat, .num, .by, nas = TRUE)
}
\arguments{
\item{.data}{a dataframe}

\item{.cat}{a catagorical column name}

\item{.num}{a numeric numerical name}

\item{.by}{categorical columns names to add a second categorical column}

\item{nas}{logical. Should missing values be considered as group on x-axis}
}
\value{
a ggplot
}
\description{
Violin plot of a numerical column by groups of a categorical column in a dataframe
}
\examples{
plot_violin(iris, Species, Sepal.Length)
}
