% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_r.R
\name{scan_r}
\alias{scan_r}
\title{R coefficient of linear models between numerical and catagorical columns in a dataframe}
\usage{
scan_r(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a data frame with paired names of categorical variables (\code{Cat}
and mnumeric ones \code{Num} and \code{r} value)
}
\description{
R coefficient of linear models between numerical and catagorical columns in a dataframe
}
\examples{
library(dplyr)
vis_r(starwars)
}
