% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{utils}
\alias{utils}
\alias{.get_madm}
\alias{.get_skewness}
\alias{.get_bins}
\alias{.get_cramerv}
\alias{.get_r}
\title{Some internals functions}
\usage{
.get_madm(.num)

.get_skewness(.num)

.get_bins(.num)

.get_cramerv(.cat1, .cat2)

.get_r(.data, .cat, .num)
}
\arguments{
\item{.num}{a numeric vector}

\item{.cat1}{a categorical vector}

\item{.cat2}{a categorical vector}

\item{.data}{a data frame}

\item{.cat}{a categorical vector}
}
\description{
Some internals functions
}
\examples{
.get_madm(rnorm(100))
.get_skewness(rnorm(100))
.get_bins(rnorm(100))
#todo
#todo
}
