% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{utils}
\alias{utils}
\alias{.get_madm}
\alias{.get_skewness}
\alias{.get_bins}
\title{Some internals functions}
\usage{
.get_madm(.num)

.get_skewness(.num)

.get_bins(.num)
}
\arguments{
\item{.num}{a numeric vector}
}
\description{
Some internals functions
}
\examples{
.get_madm(rnorm(100))
.get_skewness(rnorm(100))
.get_bins(rnorm(100))
}
