% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_cramerv.R
\name{vis_cramerv}
\alias{vis_cramerv}
\title{Plot bias corrected Cramer's V between categorical columns of a dataframe}
\usage{
vis_cramerv(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a ggplot
}
\description{
Plot bias corrected Cramer's V between categorical columns of a dataframe
}
\examples{
library(dplyr)
vis_cramerv(starwars)
}
