% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_groups.R
\name{vis_groups}
\alias{vis_groups}
\title{Ggplots of groups frequency in all categorical variables in a dataframe}
\usage{
vis_groups(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{
a list of ggplots
}
\description{
Ggplots of groups frequency in all categorical variables in a dataframe
}
\examples{
vis_groups(dplyr::starwars)
}
