% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_groups.R
\name{vis_groups}
\alias{vis_groups}
\title{Bar plots of all categorical columns in a dataframe}
\usage{
vis_groups(.data, .by)
}
\arguments{
\item{.data}{a dataframe}

\item{.by}{(optional) unquoted name of a categorical column}
}
\value{
a list of ggplots
}
\description{
Bar plots of all categorical columns in a dataframe
}
\examples{
library(dplyr)
vis_groups(starwars)
}
