% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_num.R
\name{vis_num}
\alias{vis_num}
\title{Get and Visualise distribution of a numeric vector}
\usage{
vis_num(.num, .labx = "", .bins = .get_bins(.num))
}
\arguments{
\item{.num}{a numeric vector}

\item{.labx}{label of x-axix}

\item{.bins}{bins argument for \code{geom_histogram}}
}
\value{
a ggplot object
}
\description{
Get and Visualise distribution of a numeric vector
}
\examples{
vis_num(rnorm(100))
}
