% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_numerics.R
\name{vis_numerics}
\alias{vis_numerics}
\title{Distribution of all numerical columns in a dataframe}
\usage{
vis_numerics(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a list of ggplot objects
}
\description{
Distribution of all numerical columns in a dataframe
}
\examples{
vis_numerics(iris)
}
