% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_r.R
\name{vis_r}
\alias{vis_r}
\title{Plot linear R coefficient of linear models between numeric and catagorical variables}
\usage{
vis_r(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a gpplot
}
\description{
Plot linear R coefficient of linear models between numeric and catagorical variables
}
\examples{
vis_r(dplyr::starwars)
}
