% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watch_tits}
\alias{watch_tits}
\title{Coordination of care by breeders and helpers in the cooperatively breeding long-tailed tit}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7950 rows and 21 columns.
}
\source{
Chay et al. (2022) "Coordination of care by breeders and helpers in the cooperatively breeding long-tailed tit, Aegithalos caudatus". Data available at Dryad: https://datadryad.org/dataset/doi:10.5061/dryad.mkkwh712k
}
\usage{
watch_tits
}
\description{
\itemize{
\item \code{id}: Unique identifier for each provisioning watch
\item \code{feeds}: Number of feeds by each individual per watch
\item \code{data_type}: Whether the data was taken directly from field observation (observed) or from null model randomization (expected)
\item \code{alt_feeds}: Number of alternated visits
\item \code{percent_alt}: Percentage of each carer's visits which were alternated
\item \code{sync_feeds}: Number of synchronized visits (2-minute interval)
\item \code{percent_sync2mins}: Percentage of each carer's visits which were synchronized (2-minute interval)
\item \code{carer_status}: Factor designating whether a carer was a breeding female (female), breeding male (male) or non-breeding helper (helper)
\item \code{carer_id}: The unique identifier for each carer
\item \code{total_feed_rate}: The number of feeds performed by all carers per hour
\item \code{brood_size}: The number of chicks in the brood
\item \code{carer_number}: The number of carers observed provisioning during a given watch
\item \code{watch_duration_mins}: The time, in minutes, between the first and last recorded feed during a given watch
\item \code{watch_time}: Number of hours since the start of the day 00:00 the watch was started
\item \code{brood_age}: The number of days, since recorded hatching, that a watch was performed
\item \code{julian_hatch_date}: The number of days since March 1 each year that the eggs hatched
\item \code{a_max}: The total percentage of visits during a given watch which could theoretically be alternated (or synchronized). Formula as follows: 100 - ((Number of feeds by highest feed rate carer - Number of feeds by all remaining carers)/Total number of feeds)*100
\item \code{year}: Unique identifier for each year the watch was performed during
\item \code{nest}: Unique identifier for each nest
\item \code{individual_feed_rate}: The number of feeds performed by each carer per hour
\item \code{row_ref}: Row reference 1-7950
}
}
\keyword{datasets}
